+++ src/navigation/ScreensEnum.js
@@ -1,0 +1,8 @@
+const SCREEN = {
+    ERROR: 'ErrorScreen',
+    EXAMPLE: 'ExampleScreen',
+    EXAMPLE2: 'ExampleScreen2',
+    MAIN: 'MainScreen',
+}
+
+export default SCREEN