+++ src/services/ErrorReportingService.js
@@ -1,21 +1,28 @@
+// Functions related to reporting errors
+
 // Third-party dependencies
 import * as Sentry from '@sentry/react-native'
 
-export function safeReportError(error, errorInfo) {
+function safeReportError(error, errorInfo) {
     try {
         console.log(error)
 
         Sentry.withScope((scope) => {
             if (errorInfo) {
-              scope.setExtras(errorInfo)
+                scope.setExtras(errorInfo)
             }
             Sentry.captureException(error)
-          })
+        })
     } catch (err) {
         console.log('Error reporting error', err)
     }
 }
 
-export function simulateCrash() {
+function simulateCrash() {
     throw 'This error was generated to test error handling and can be safely ignored'
+}
+
+export {
+    safeReportError,
+    simulateCrash,
 }