+++ src/services/errors/FetchServiceUnauthorizedClientError.js
@@ -1,0 +1,14 @@
+class FetchServiceUnauthorizedClientError extends Error {
+
+    constructor(...params) {
+        super(...params)
+
+        if (Error.captureStackTrace) {
+            Error.captureStackTrace(this, FetchServiceUnauthorizedClientError)
+        }
+
+        this.name = 'FetchServiceUnauthorizedClientError'
+    }
+}
+
+export default FetchServiceUnauthorizedClientError